+++ neurolang/relational_algebra/__init__.py
@@ -1,0 +1,2 @@
+from .relational_algebra import *
+from .optimisers import *
